Create a set of interview handouts for a SME

Creates two MS Word formatted documents. One is an `answers` document that contains the answers to the calibration questions, the other does not contain answers and is intended to be a visual reference (and possible take away) for the SME.

make_handouts(sme, questions, output_dir = getwd())

Arguments

sme
Name of the SME.

questions
questions object

output_dir
Directory to place output.

Value

Output of render.

Examples

# NOT RUN {
questions <- read_questions()
make_handouts("Sally Expert", questions)
# }